screengrab require optimizations/fixes
