+++ screengrab/lib/screengrab/commands_generator.rb
@@ -1,5 +1,13 @@
 require 'commander'
+
 require 'fastlane/version'
+require 'fastlane/fastlane_folder'
+require 'fastlane_core/configuration/configuration'
+require_relative 'android_environment'
+require_relative 'dependency_checker'
+require_relative 'runner'
+require_relative 'options'
+require_relative 'module'
 
 HighLine.track_eof = false
