+++ screengrab/lib/screengrab/runner.rb
@@ -1,3 +1,7 @@
+require 'fastlane_core/print_table'
+require 'fastlane_core/command_executor'
+require_relative 'reports_generator'
+require_relative 'module'
 
 module Screengrab
   class Runner
